\name{dna.identity}
\Rdversion{1.1}
\alias{dna.identity}
\docType{data}
\title{dna.identity}
\description{DNA identity substitution matrix.}
\usage{dna.identity}
\format{
  The format is:
 num [1:5, 1:5] 0 0 0 0 0 0 1 0 0 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:5] "*" "A" "T" "G" ...
  ..$ : chr [1:5] "*" "A" "T" "G" ...
}
\details{
}
\source{
}
\references{
}