\name{read.fasta}
\Rdversion{1.1}
\alias{read.fasta}
\title{read.fasta}
\description{Read sequences in FASTA format into a named character vector}
\usage{
read.fasta(infile)
}
\arguments{
  \item{infile}{Name of the sequence file}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}