\name{sublogo-package}
\Rdversion{1.1}
\alias{sublogo-package}
\docType{package}
\title{Sublogo dendrogram plots}
\description{Visualize correlation in biological sequence data using sublogo dendrogram plots.}
\details{
\tabular{ll}{Package: \tab sublogo\cr
Type: \tab Package\cr
Title: \tab Sublogo dendrogram plots\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-19\cr
Author: \tab Toby Dylan Hocking\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@etu.upmc.fr>\cr
Depends: \tab R (>= 2.3.0), grImport, gridBase\cr
##Suggests: \tab seqinr\cr
Description: \tab Visualize correlation in biological sequence data using sublogo dendrogram plots.\cr
URL: \tab http://sublogo.r-forge.r-project.org\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}