\name{sublogo}
\Rdversion{1.1}
\alias{sublogo}
\title{sublogo}
\description{Draw a sublogo dendrogram for a sequence motif.}
\usage{
sublogo(seqs, mat = NULL, ...)
}
\arguments{
  \item{seqs}{Character vector of DNA or protein sequences (optionally named with sequence labels).}
  \item{mat}{Substitution matrix passed to seqs.to.mat.}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}
\examples{}