\name{make.logo.ps}
\Rdversion{1.1}
\alias{make.logo.ps}
\title{make.logo.ps}
\description{Create a logo using weblogo, then read it in using grImport}
\usage{
make.logo.ps(helices, psbase)
}
\arguments{
  \item{helices}{Sequences to plot in the logo}
  \item{psbase}{Base filename for the logo postscript and xml files, should be the full path name except for the trailing .ps}
}
\details{
}
\value{Grid picture grob as read using readPicture}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}